splash removed to get into index
