+++ src/pages/index.tsx
@@ -1,26 +1,40 @@
 import Head from "next/head";
 import { PrismaClient } from "@prisma/client";
+import { LargeButton } from "@/components/LargeButton";
+import Link from "next/link";
+import Image from "next/image";
 
 export const prisma = new PrismaClient();
 
-export default function Home() {
-	return (
-		<>
-			<Head>
-				<title>Create Next App</title>
-				<meta
-					name="description"
-					content="Generated by create next app"
-				/>
-				<meta
-					name="viewport"
-					content="width=device-width, initial-scale=1"
-				/>
-				<link rel="icon" href="/favicon.ico" />
-			</Head>
-			<h1 className="text-red-600 text-4xl font-bold flex justify-center">
-				Landing Page
-			</h1>
-		</>
-	);
+export default function Splash() {
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div className="flex flex-col pt-12 justify-center items-center">
+        <Image
+          src="/images/listfull-logo.png"
+          alt=""
+          width={218}
+          height={186}
+        />
+      </div>
+      <h1 className=" text-text-typo text-splash px-8">
+        HI THERE! WELCOME TO SHOPPING HELPER
+      </h1>
+      <div className="px-8 pt-10">
+        <Link href={"/login"}>
+          <LargeButton
+            variant={"primary"}
+            label={"get started"}
+            disabled={false}
+          ></LargeButton>
+        </Link>
+      </div>
+    </>
+  );
 }
