--- src/pages/splash/index.tsx
@@ -1,37 +1,0 @@
-import { LargeButton } from "@/components/LargeButton";
-import Head from "next/head";
-import Link from "next/link";
-import Image from "next/image";
-
-export default function Splash() {
-  return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <div className="flex flex-col pt-12 justify-center items-center">
-        <Image
-          src="/images/listfull-logo.png"
-          alt=""
-          width={218}
-          height={186}
-        />
-      </div>
-      <h1 className=" text-text-typo text-splash px-8">
-        HI THERE! WELCOME TO SHOPPING HELPER
-      </h1>
-      <div className="px-8 pt-10">
-        <Link href={"/login"}>
-          <LargeButton
-            variant={"primary"}
-            label={"get started"}
-            disabled={false}
-          ></LargeButton>
-        </Link>
-      </div>
-    </>
-  );
-}
